`TileEntityInternalMultiblock.onNeighborChange` no longer casts the position difference to `long`.

+++ src/main/java/mekanism/common/tile/prefab/TileEntityInternalMultiblock.java
@@ -69,9 +69,9 @@ public class TileEntityInternalMultiblock extends TileEntityMekanism implements 
         //If the neighbor change happened to a block outside a multiblock, and it isn't a block that could be part of the multiblock
         //Note: This is necessary to eliminate unnecessary structure recheck
         // caused by redstone optimization mods like Alternate Current reporting the different neighborPos from Vanilla implementation.
-        long dX = Math.abs((long)neighborPos.getX() - worldPosition.getX());
-        long dY = Math.abs((long)neighborPos.getY() - worldPosition.getY());
-        long dZ = Math.abs((long)neighborPos.getZ() - worldPosition.getZ());
+        int dX = Math.abs(neighborPos.getX() - worldPosition.getX());
+        int dY = Math.abs(neighborPos.getY() - worldPosition.getY());
+        int dZ = Math.abs(neighborPos.getZ() - worldPosition.getZ());
         if (Math.max(Math.max(dX, dY), dZ) > 1) return;
         //And we are not already an internal part of the structure, or we are changing an internal part to air
         // then we mark the structure as needing to be re-validated
